split base trait into a separate file, load as mod
